% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{cuml_umap_dbs_clust}
\alias{cuml_umap_dbs_clust}
\title{Sequential UMAP 2d embedding and simple 2d space partitioning clustering (CUML accelerated version).}
\usage{
cuml_umap_dbs_clust(
  data,
  maxevts,
  maxlvl,
  minpts,
  clust_options,
  ret_model = F,
  myqueue = NULL,
  verbose = F
)
}
\arguments{
\item{data}{a data matrix.}

\item{maxevts}{max number of random points to apply UMAP.}

\item{maxlvl}{max sequential levels to process.}

\item{minpts}{min cluster size.}

\item{clust_options}{list; 2d clustering options (see Details).}

\item{ret_model}{logical; should clustering model be exported.}

\item{myqueue}{shinyQueue object. Multithreaded code use \link[future]{future} package. This object is used to send log messages to the base worker.}

\item{verbose}{logical.}
}
\value{
A vector of cluster numbers with length = nrow(data)
}
\description{
Sequential UMAP 2d embedding and simple 2d space partitioning clustering (CUML accelerated version).
}
\details{
clust_options must be a named list containing:
 method: one of "dbscan", "hdbscan", "kdbscan", "sdbscan"
 mineps: min eps value passed to dbscan
 mindens: min density passed to kdbscan
 bw: bw passed to sdbscan
 nbins: min number of bins of bw width
 mvpratio: max valley/peak ratio)
}
