% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{kdbscan}
\alias{kdbscan}
\title{Kernel density based 2d clustering.}
\usage{
kdbscan(
  x,
  minpts = 100,
  mindens = 0.001,
  maxlvl = 100,
  alfa = 0.05,
  theta = 5,
  ret_model = FALSE
)
}
\arguments{
\item{x}{a data matrix.}

\item{minpts}{min cluster size.}

\item{mindens}{density cutoff.}

\item{maxlvl}{max sequential level of space partitioning.}

\item{alfa}{numeric. Exclude alfa portion of each extremity before search for a cut point in density curve.}

\item{theta}{integer. Angle of rotation in each step.}

\item{ret_model}{logical.}
}
\value{
A vector with cluster numbers. Length = nrow(x)
}
\description{
Kernel density based 2d clustering.
}
