% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{plot.s2dcluster}
\alias{plot.s2dcluster}
\title{Plot s2dcluster model}
\usage{
\method{plot}{s2dcluster}(x, data = NULL, cex = 2, ...)
}
\arguments{
\item{x}{s2dcluster model}

\item{data}{a data matrix. Original data saved in model used if NULL.}

\item{cex}{point size,}

\item{...}{arguments passed to plot.xy}
}
\description{
Plot s2dcluster model
}
