% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{plot.sumapc}
\alias{plot.sumapc}
\title{Plot sumapc model data.}
\usage{
\method{plot}{sumapc}(x, data = NULL, ...)
}
\arguments{
\item{x}{sumapc}

\item{data}{a data matrix. If NULL uses model data.}

\item{...}{arguments passed to plot.s2dcluster}
}
\description{
Plot sumapc model data.
}
