% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{plot.UMAP_s2dcluster}
\alias{plot.UMAP_s2dcluster}
\title{Plot UMAP_s2dcluster model data.}
\usage{
\method{plot}{UMAP_s2dcluster}(x, data = NULL, ...)
}
\arguments{
\item{x}{UMAP_s2dcluster}

\item{data}{a data matrix. If NULL uses model data.}

\item{...}{arguments passed to plot.s2dcluster}
}
\description{
Plot UMAP_s2dcluster model data.
}
