% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{predict.s2dcluster}
\alias{predict.s2dcluster}
\title{Predict clusters based on s2dcluster model.}
\usage{
\method{predict}{s2dcluster}(object, newdata, ...)
}
\arguments{
\item{object}{s2dcluster model.}

\item{newdata}{a data matrix.}

\item{...}{NOT USED}
}
\value{
A vector of cluster numbers.
}
\description{
Predict clusters based on s2dcluster model.
}
