% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{predict.sumapc}
\alias{predict.sumapc}
\title{Predict clusters based on a sumapc model}
\usage{
\method{predict}{sumapc}(
  object,
  newdata,
  batch = Inf,
  multi_thread = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{sumapc model.}

\item{newdata}{a data matrix.}

\item{batch}{integer. Split newdata in small batches reducing memory usage.}

\item{multi_thread}{logical.}

\item{verbose}{logical.}

\item{...}{NOT USED.}
}
\value{
A vector of cluster numbers with length = nrow(data)
}
\description{
Predict clusters based on a sumapc model
}
