% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{predict.UMAP_s2dcluster}
\alias{predict.UMAP_s2dcluster}
\title{Predict clusters based on a UMAP_s2dcluster model}
\usage{
\method{predict}{UMAP_s2dcluster}(object, newdata, multi_thread = T, verbose = F, ...)
}
\arguments{
\item{object}{UMAP_s2dcluster model.}

\item{newdata}{a data matrix.}

\item{multi_thread}{logical.}

\item{verbose}{logical.}

\item{...}{NOT USED.}
}
\value{
A vector of cluster numbers with length = nrow(data)
}
\description{
Predict clusters based on a UMAP_s2dcluster model
}
