% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{print.s2dcluster}
\alias{print.s2dcluster}
\title{Print s2dcluster information.}
\usage{
\method{print}{s2dcluster}(x, ...)
}
\arguments{
\item{x}{s2dcluster model}

\item{...}{NOT USED}
}
\description{
Print s2dcluster information.
}
