% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{print.sumapc}
\alias{print.sumapc}
\title{Print sumapc model information}
\usage{
\method{print}{sumapc}(x, ...)
}
\arguments{
\item{x}{sumapc model}

\item{...}{NOT USED}
}
\description{
Print sumapc model information
}
