% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{save_model}
\alias{save_model}
\title{Save UMAP_s2dcluster model to file}
\usage{
save_model(res, file)
}
\arguments{
\item{res}{UMAP_s2dcluster model.}

\item{file}{filename.}
}
\description{
Save UMAP_s2dcluster model to file
}
