% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumapc.R
\name{sdbscan}
\alias{sdbscan}
\title{Simple density based 2d clustering.}
\usage{
sdbscan(
  x,
  minpts = 100,
  maxlvl = 100,
  alfa = 0.05,
  bw = 0.25,
  nbins = 1,
  theta = 5,
  mvpratio = 0.5,
  ret_model = F,
  plotcuts = F
)
}
\arguments{
\item{x}{a data matrix.}

\item{minpts}{min cluster size.}

\item{maxlvl}{max sequential level of space partitioning.}

\item{alfa}{numeric. Exclude alfa portion of each extremity before search for a cut point in density curve.}

\item{bw}{bin width.}

\item{nbins}{min number of bins to search.}

\item{theta}{integer. Angle of rotation in each step.}

\item{mvpratio}{max valley/peaks value to allow cut.}

\item{ret_model}{logical.}

\item{plotcuts}{logical; plot data with histograms and cut point.}
}
\value{
A vector with cluster numbers. Length = nrow(x)
}
\description{
Simple density based 2d clustering.
}
